requirements.txt
amqp==5.2.0
asgiref==3.7.2
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-filter==24.1
django-import-export==3.3.8
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.6.1
et-xmlfile==1.1.0
eventlet==0.36.1
greenlet==3.0.3
httplib2==0.22.0
idna==3.7
kombu==5.3.7
Markdown==3.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.3.0
prompt-toolkit==3.0.43
PyJWT==2.8.0
pyparsing==3.1.2
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
